+++ source/framework/vx_target.c
@@ -137,7 +137,7 @@ static void tivxTargetFreeHandle(tivx_target *target_handle)
 static vx_status tivxTargetDequeueObjDesc(tivx_target target, uint16_t *obj_desc_id, uint32_t timeout)
 {
     vx_status status;
-    uint32_t value = TIVX_OBJ_DESC_INVALID;
+    uintptr_t value = TIVX_OBJ_DESC_INVALID;
 
     *obj_desc_id = TIVX_OBJ_DESC_INVALID;
 
@@ -1125,7 +1125,7 @@ vx_status tivxTargetQueueObjDesc(vx_enum target_id, uint16_t obj_desc_id)
     if(target!=NULL)
     {
         status = tivxQueuePut(&target->job_queue_handle,
-                obj_desc_id, TIVX_EVENT_TIMEOUT_NO_WAIT);
+                (uintptr_t)obj_desc_id, TIVX_EVENT_TIMEOUT_NO_WAIT);
     }
 
     return status;
