+++ source/platform/os/win32/tivx_queue.c
@@ -69,7 +69,7 @@
 
 
 vx_status tivxQueueCreate(
-    tivx_queue *queue, uint32_t max_elements, uint32_t *queue_memory,
+    tivx_queue *queue, uint32_t max_elements, uintptr_t *queue_memory,
     uint32_t flags)
 {
     vx_status status = VX_FAILURE;
@@ -159,7 +159,7 @@ vx_status tivxQueueDelete(tivx_queue *queue)
     return (status);
 }
 
-vx_status tivxQueuePut(tivx_queue *queue, uint32_t data, uint32_t timeout)
+vx_status tivxQueuePut(tivx_queue *queue, uintptr_t data, uint32_t timeout)
 {
     vx_status status = VX_FAILURE;
     volatile vx_bool do_break = vx_false_e;
@@ -250,7 +250,7 @@ vx_status tivxQueuePut(tivx_queue *queue, uint32_t data, uint32_t timeout)
     return (status);
 }
 
-vx_status tivxQueueGet(tivx_queue *queue, uint32_t *data, uint32_t timeout)
+vx_status tivxQueueGet(tivx_queue *queue, uintptr_t *data, uint32_t timeout)
 {
     vx_status status = VX_FAILURE;/* init status to error */
     volatile vx_bool do_break = vx_false_e;
